# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Medical Image Analysis Platform",
    "description": "A production-ready API for medical image processing with real-time updates",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/login/token": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login For Access Token",
        "operationId": "login_for_access_token_api_v1_login_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analysis/start": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Start Analysis",
        "description": "Start AI analysis for an uploaded image.\n...",
        "operationId": "start_analysis_api_v1_analysis_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/{analysis_id}": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "Get Analysis Result",
        "description": "Get analysis result by ID.\n\nThis endpoint provides the current status and results of an analysis.\nClients can poll this endpoint or use WebSocket for real-time updates.\n\nArgs:\n    analysis_id: Unique analysis identifier\n    db: Database session\n    \nReturns:\n    AnalysisResponse with current status and results",
        "operationId": "get_analysis_result_api_v1_analysis__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Analysis"
        ],
        "summary": "Delete Analysis",
        "description": "Delete analysis record.\n\nThis endpoint allows deletion of any analysis regardless of status.\nFor running analyses, it will cancel them first, then delete.\nFor completed analyses, it will just delete the record.",
        "operationId": "delete_analysis_api_v1_analysis__analysis_id__delete",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "List Analyses",
        "description": "List analysis results with optional filtering.\n\nArgs:\n    skip: Number of records to skip (pagination)\n    limit: Maximum number of records to return\n    status: Filter by analysis status\n    user_id: Filter by user who requested analysis\n    db: Database session\n    \nReturns:\n    List of AnalysisResponse objects",
        "operationId": "list_analyses_api_v1_analysis__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AnalysisStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisResponse"
                  },
                  "title": "Response List Analyses Api V1 Analysis  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/models/": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "List Ai Models",
        "description": "List available AI models.\n\nArgs:\n    active_only: Whether to return only active models\n    db: Database session\n    \nReturns:\n    List of AI model information",
        "operationId": "list_ai_models_api_v1_analysis_models__get",
        "parameters": [
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response List Ai Models Api V1 Analysis Models  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisRequest": {
        "properties": {
          "image_id": {
            "type": "string",
            "title": "Image Id",
            "description": "ID of image to analyze"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id",
            "description": "ID of AI model to use"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "description": "User requesting the analysis"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority",
            "description": "Analysis priority: low, normal, high",
            "default": "normal"
          }
        },
        "type": "object",
        "required": [
          "image_id",
          "model_id"
        ],
        "title": "AnalysisRequest",
        "description": "Request schema for starting AI analysis."
      },
      "AnalysisResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "ai_model_id": {
            "type": "string",
            "title": "Ai Model Id"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "progress_percentage": {
            "type": "number",
            "maximum": 100,
            "minimum": 0,
            "title": "Progress Percentage"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1,
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "results": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Results"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "processing_time_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Time Seconds"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "image_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Filename"
          },
          "analysis_model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Model Name"
          },
          "analysis_model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Model Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "image_id",
          "ai_model_id",
          "status",
          "progress_percentage",
          "created_at",
          "updated_at"
        ],
        "title": "AnalysisResponse",
        "description": "Complete analysis result response schema."
      },
      "AnalysisStartResponse": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id",
            "description": "Unique analysis identifier for tracking"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AnalysisStatus"
              }
            ],
            "description": "Initial analysis status"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Success message"
          },
          "estimated_completion_time": {
            "type": "integer",
            "title": "Estimated Completion Time",
            "description": "Estimated completion time in seconds"
          },
          "websocket_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Websocket Url",
            "description": "WebSocket URL for real-time updates"
          }
        },
        "type": "object",
        "required": [
          "analysis_id",
          "status",
          "message",
          "estimated_completion_time"
        ],
        "title": "AnalysisStartResponse",
        "description": "Response schema for analysis start request."
      },
      "AnalysisStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ANALYZING",
          "COMPLETE",
          "FAILED",
          "CANCELLED"
        ],
        "title": "AnalysisStatus",
        "description": "Analysis status enumeration."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TokenRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "TokenRequest"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/api/v1/login/token"
          }
        }
      }
    }
  }
}